requirements.txt
Flask==2.1.2
Werkzeug==2.1.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
pycryptodome==3.15.0
Flask-Classful==0.14.2
requests==2.22.0
rsa==4.9

